Fiz um esquema onde quando o usuário carrega a pagina de login, as variáveis globais dos inputs são limpos
