+++ styles/globalStyles.js
@@ -61,7 +61,21 @@ const globalStyles = StyleSheet.create({
         marginLeft: "20%",
         tintColor: '#fff',
     },
-    
+    inputBoxPass: {
+        flexDirection: 'row',
+        alignItems: 'center',
+    },
+    eyeStyle: {
+        position: 'absolute',
+        marginRight: 10,
+        right: 0,
+        opacity: 0.6
+    },
+    eyeImageStyle: {
+        marginRight: 2,
+        height: 18,
+        width: 18
+    },
   
 })
 export default globalStyles;
